add: bibtex

+++ cpp/main.cpp
@@ -8,7 +8,7 @@ using namespace std;
   #include "print.hpp"
 #endif
 
-#define rep(i, n) for(int i = 0; i < (int)(n); i++)
+#define rep(i, n) for(ll i = 0; i < (ll)(n); i++)
 #define ALL(x) (x).begin(), (x).end()
 #define RALL(x) (x).rbegin(), (x).rend()
 #define MOD 1000000007
